refactor(price-fetcher): pure 헬퍼를 사이드이펙트 없는 별도 모듈로 분리 (Codex P2, #420 follow-up)
Codex 리뷰 (PR #429 review 4660255262): 테스트가 `../price-fetcher` 를
직접 import 하면 top-level Prisma client (`src/lib/prisma.ts`) 가 로드됨.
Prisma generate / DATABASE_URL 없는 CI 환경에서 모듈 import 자체가 실패.

## 수정
- `mergeCrossTickersIntoMeta` + `TickerMetaValue` 타입을 `price-fetcher-utils.ts`
  로 분리 (import 없음, 100% pure)
- 본체 `price-fetcher.ts` 는 utils 를 재수출 (기존 import 경로 하위호환)
- 테스트를 `price-fetcher-utils.test.ts` 로 rename + import 도 utils 모듈에서

## 검증
- 격리 실행 (`npm run test -- --run price-fetcher-utils.test.ts`) → 5 pass,
  Prisma import chain 없이 통과
- 전체 470 pass

Closes #420 (follow-up, PR #429)
